refactor mcp go dto fields
